Moved builtin_costs libfunc into the gas modules
