+++ crates/cairo-lang-sierra-gas/src/gas_info.rs
@@ -1,6 +1,6 @@
 use std::fmt::Display;
 
-use cairo_lang_sierra::extensions::builtin_cost::CostTokenType;
+use cairo_lang_sierra::extensions::gas::CostTokenType;
 use cairo_lang_sierra::ids::FunctionId;
 use cairo_lang_sierra::program::StatementIdx;
 use cairo_lang_utils::ordered_hash_map::OrderedHashMap;
