+++ crates/cairo-lang-sierra-gas/src/generate_equations.rs
@@ -1,4 +1,4 @@
-use cairo_lang_sierra::extensions::builtin_cost::CostTokenType;
+use cairo_lang_sierra::extensions::gas::CostTokenType;
 use cairo_lang_sierra::ids::ConcreteLibfuncId;
 use cairo_lang_sierra::program::{Program, StatementIdx};
 use cairo_lang_utils::collection_arithmetics::{add_maps, sub_maps};
