+++ crates/cairo-lang-sierra-gas/src/objects.rs
@@ -1,4 +1,4 @@
-use cairo_lang_sierra::extensions::builtin_cost::CostTokenType;
+use cairo_lang_sierra::extensions::gas::CostTokenType;
 use cairo_lang_sierra::ids::ConcreteTypeId;
 use cairo_lang_sierra::program::Function;
 use cairo_lang_utils::collection_arithmetics::{add_maps, sub_maps};
