+++ crates/cairo-lang-sierra-to-casm/src/environment/gas_wallet.rs
@@ -1,6 +1,6 @@
 use std::fmt::Display;
 
-use cairo_lang_sierra::extensions::builtin_cost::CostTokenType;
+use cairo_lang_sierra::extensions::gas::CostTokenType;
 use cairo_lang_utils::collection_arithmetics::add_maps;
 use cairo_lang_utils::ordered_hash_map::OrderedHashMap;
 use thiserror::Error;
