+++ crates/cairo-lang-sierra/src/simulation/core.rs
@@ -22,7 +22,9 @@ use crate::extensions::felt252::{
 };
 use crate::extensions::felt252_dict::Felt252DictConcreteLibfunc;
 use crate::extensions::function_call::FunctionCallConcreteLibfunc;
-use crate::extensions::gas::GasConcreteLibfunc::{GetAvailableGas, RedepositGas, WithdrawGas};
+use crate::extensions::gas::GasConcreteLibfunc::{
+    BuiltinWithdrawGas, GetAvailableGas, GetBuiltinCosts, RedepositGas, WithdrawGas,
+};
 use crate::extensions::mem::MemConcreteLibfunc::{
     AllocLocal, FinalizeLocals, Rename, StoreLocal, StoreTemp,
 };
@@ -138,6 +140,9 @@ pub fn simulate<
                 vec![CoreValue::GasBuiltin(*gas_counter), CoreValue::Uint128(*gas_counter as u128)],
                 0,
             ))
+        }
+        Gas(BuiltinWithdrawGas(_) | GetBuiltinCosts(_)) => {
+            unimplemented!("Simulation of the builtin cost functionality is not implemented yet.")
         }
         BranchAlign(_) => {
             get_statement_gas_info().ok_or(LibfuncSimulationError::UnresolvedStatementGasInfo)?;
@@ -335,9 +340,6 @@ pub fn simulate<
         }
         CoreConcreteLibfunc::Poseidon(_) => {
             unimplemented!("Simulation of the Poseidon hash function is not implemented yet.");
-        }
-        CoreConcreteLibfunc::BuiltinCost(_) => {
-            unimplemented!("Simulation of the builtin cost functionality is not implemented yet.")
         }
         CoreConcreteLibfunc::StarkNet(_) => {
             unimplemented!("Simulation of the StarkNet functionalities is not implemented yet.")
